+++ src/test/java/ladder/domain/LadderTest.java
@@ -1,6 +1,5 @@
 package ladder.domain;
 
-import ladder.strategy.BridgeLinesRandomStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
